requirements.txt
pygame-ce ~= 2.5.6